+++ services/cases.js
@@ -8,7 +8,7 @@ const Check = require('../helpers/rolecheck')
 const Notif = require('../helpers/notification')
 const Filter = require('../helpers/filter/casefilter')
 const CloseContact = require('../models/CloseContact')
-const { doUpdateEmbeddedClosecontactDoc } = require('../helpers/cases/setter')
+const { doUpdateEmbeddedClosecontactDoc } = require('../helpers/cases/setters')
 const { sqlCondition, excellOutput } = require('../helpers/filter/exportfilter')
 const { CRITERIA, WHERE_GLOBAL } = require('../helpers/constant')
 const moment = require('moment')
